Moved all exceptions to lowest level.
This is typically bad code practice, as it makes it near impossible for
other devs to add in custom behavior for catching exceptions. In almost
every case ever, the custom behavior is going to be exactly the same as
the behavior I have specified in this PR. Over all it makes it so devs
using this lib can write cleaner code.